refactor: create subfolders inside Button component
