+++ src/components/SharedComponents/Button/ButtonAsLink/ButtonAsLink.test.tsx
@@ -1,6 +1,5 @@
-import renderWithTheme from "../../../__tests__/__helpers__/renderWithTheme";
-
-import ButtonAsLink from ".";
+import renderWithTheme from "../../../../__tests__/__helpers__/renderWithTheme";
+import ButtonAsLink from "..";
 
 describe("ButtonAsLink", () => {
   it("renders a button as a link", () => {
