+++ src/components/SharedComponents/Button/ButtonAsLink/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./ButtonAsLink";
+export type { ButtonAsLinkProps } from "./ButtonAsLink";
